Add contextual leveled logger (-debug) + fixes.

+++ src/main/java/com/karalabe/iris/schemes/BroadcastScheme.java
@@ -8,6 +8,7 @@ package com.karalabe.iris.schemes;
 import com.karalabe.iris.ServiceHandler;
 import com.karalabe.iris.ServiceLimits;
 import com.karalabe.iris.common.BoundedThreadPool;
+import com.karalabe.iris.common.ContextualLogger;
 import com.karalabe.iris.protocol.RelayProtocol;
 
 import java.io.IOException;
@@ -16,13 +17,22 @@ import java.io.IOException;
 public class BroadcastScheme {
     private final RelayProtocol     protocol; // Network connection implementing the relay protocol
     private final ServiceHandler    handler;  // Callback handler for processing inbound broadcasts
+    private final ServiceLimits     limits;   // Service handler resource consumption allowance
     private final BoundedThreadPool workers;  // Thread pool for limiting the concurrent processing
+    private final ContextualLogger  logger;   // Logger with connection id injected
 
     // Constructs a broadcast scheme implementation.
-    public BroadcastScheme(final RelayProtocol protocol, final ServiceHandler handler, final ServiceLimits limits) {
+    public BroadcastScheme(final RelayProtocol protocol, final ServiceHandler handler, final ServiceLimits limits, final ContextualLogger logger) {
         this.protocol = protocol;
         this.handler = handler;
-        this.workers = new BoundedThreadPool(limits.broadcastThreads, limits.broadcastMemory);
+        this.limits = limits;
+        this.logger = logger;
+
+        if (limits != null) {
+            this.workers = new BoundedThreadPool(limits.broadcastThreads, limits.broadcastMemory);
+        } else {
+            this.workers = null;
+        }
     }
 
     // Relays a broadcast operation to the local Iris node.
@@ -32,11 +42,19 @@ public class BroadcastScheme {
 
     // Schedules an application broadcast message for the service handler to process.
     public void handleBroadcast(final byte[] message) {
-        workers.schedule(() -> handler.handleBroadcast(message), message.length);
+        if (!workers.schedule(() -> handler.handleBroadcast(message), message.length)) {
+            logger.loadContext();
+            logger.error("Broadcast exceeded memory allowance",
+                         "limit", String.valueOf(limits.broadcastMemory),
+                         "size", String.valueOf(message.length));
+            logger.unloadContext();
+        }
     }
 
     // Terminates the broadcast primitive.
     public void close() throws InterruptedException {
-        workers.terminate(true);
+        if (workers != null) {
+            workers.terminate(true);
+        }
     }
 }
