+++ src/main/java/com/karalabe/iris/schemes/RequestScheme.java
@@ -8,6 +8,7 @@ package com.karalabe.iris.schemes;
 import com.karalabe.iris.ServiceHandler;
 import com.karalabe.iris.ServiceLimits;
 import com.karalabe.iris.common.BoundedThreadPool;
+import com.karalabe.iris.common.ContextualLogger;
 import com.karalabe.iris.exceptions.RemoteException;
 import com.karalabe.iris.exceptions.TimeoutException;
 import com.karalabe.iris.protocol.RelayProtocol;
@@ -27,16 +28,25 @@ public class RequestScheme {
 
     private final RelayProtocol     protocol; // Network connection implementing the relay protocol
     private final ServiceHandler    handler;  // Callback handler for processing inbound requests
+    private final ServiceLimits     limits;   // Service handler resource consumption allowance
     private final BoundedThreadPool workers;  // Thread pool for limiting the concurrent processing
+    private final ContextualLogger  logger;   // Logger with connection id injected
 
     private final AtomicLong                nextId  = new AtomicLong();             // Unique identifier for the next request
     private final Map<Long, PendingRequest> pending = new ConcurrentHashMap<>(128); // Result objects for pending requests
 
     // Constructs a request/reply scheme implementation.
-    public RequestScheme(final RelayProtocol protocol, final ServiceHandler handler, final ServiceLimits limits) {
+    public RequestScheme(final RelayProtocol protocol, final ServiceHandler handler, final ServiceLimits limits, final ContextualLogger logger) {
         this.protocol = protocol;
         this.handler = handler;
-        this.workers = new BoundedThreadPool(limits.requestThreads, limits.requestMemory);
+        this.limits = limits;
+        this.logger = logger;
+
+        if (limits != null) {
+            this.workers = new BoundedThreadPool(limits.requestThreads, limits.requestMemory);
+        } else {
+            this.workers = null;
+        }
     }
 
     // Relays a request to the local Iris node, waits for a reply or timeout and returns it.
@@ -70,7 +80,15 @@ public class RequestScheme {
 
     // Schedules an application request for the service handler to process.
     public void handleRequest(final long id, final byte[] request, final long timeout) {
-        workers.schedule(() -> {
+        final long start = System.nanoTime();
+        if (!workers.schedule(() -> {
+            // Ensure that expired tasks get dropped instead of executed
+            if (((System.nanoTime() - start) / 1_000_000) >= timeout) {
+                logger.error("Dumping expired scheduled request",
+                             "timeout", String.valueOf(timeout));
+                return;
+            }
+
             byte[] response = null;
             String error = null;
 
@@ -84,9 +102,15 @@ public class RequestScheme {
             try {
                 reply(id, response, error);
             } catch (IOException e) {
-                e.printStackTrace();
+                logger.error("Failed to send reply", "reason", e.getMessage());
             }
-        }, request.length, (int) timeout);
+        }, request.length)) {
+            logger.loadContext();
+            logger.error("Request exceeded memory allowance",
+                         "limit", String.valueOf(limits.requestMemory),
+                         "size", String.valueOf(request.length));
+            logger.unloadContext();
+        }
     }
 
     // Relays a reply to a request to the local Iris node.
@@ -114,6 +138,8 @@ public class RequestScheme {
 
     // Terminates the request/reply primitive.
     public void close() throws InterruptedException {
-        workers.terminate(true);
+        if (workers != null) {
+            workers.terminate(true);
+        }
     }
 }
